# api-docs.json
{"openapi":"3.0.3","info":{"title":"黑马点评接口文档","description":"黑马点评项目接口文档","version":"1.0"},"servers":[{"url":"http://localhost:8081","description":"Inferred Url"}],"tags":[{"name":"blog-controller","extensions":{"x-order":"2147483647"}},{"name":"follow-controller","extensions":{"x-order":"2147483647"}},{"name":"shop-controller","extensions":{"x-order":"2147483647"}},{"name":"shop-type-controller","extensions":{"x-order":"2147483647"}},{"name":"upload-controller","extensions":{"x-order":"2147483647"}},{"name":"user-controller","extensions":{"x-order":"2147483647"}},{"name":"voucher-controller","extensions":{"x-order":"2147483647"}},{"name":"voucher-order-controller","extensions":{"x-order":"2147483647"}}],"paths":{"/blog":{"post":{"tags":["blog-controller"],"summary":"saveBlog","operationId":"saveBlogUsingPOST","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Blog"}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Result"}}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"extensions":{"x-order":"2147483647"}}},"/blog/hot":{"get":{"tags":["blog-controller"],"summary":"queryHotBlog","operationId":"queryHotBlogUsingGET","parameters":[{"name":"current","in":"query","description":"current","required":false,"style":"form","schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Result"}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"extensions":{"x-order":"2147483647"}}},"/blog/like/{id}":{"put":{"tags":["blog-controller"],"summary":"likeBlog","operationId":"likeBlogUsingPUT","parameters":[{"name":"id","in":"path","description":"id","required":true,"style":"simple","schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Result"}}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"extensions":{"x-order":"2147483647"}}},"/blog/likes/{id}":{"get":{"tags":["blog-controller"],"summary":"queryBlogLikes","operationId":"queryBlogLikesUsingGET","parameters":[{"name":"id","in":"path","description":"id","required":true,"style":"simple","schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Result"}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"extensions":{"x-order":"2147483647"}}},"/blog/of/follow":{"get":{"tags":["blog-controller"],"summary":"queryBlogOfFollow","operationId":"queryBlogOfFollowUsingGET","parameters":[{"name":"lastId","in":"query","description":"lastId","required":true,"style":"form","schema":{"type":"integer","format":"int64"}},{"name":"offset","in":"query","description":"offset","required":false,"style":"form","schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Result"}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"extensions":{"x-order":"2147483647"}}},"/blog/of/me":{"get":{"tags":["blog-controller"],"summary":"queryMyBlog","operationId":"queryMyBlogUsingGET","parameters":[{"name":"current","in":"query","description":"current","required":false,"style":"form","schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Result"}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"extensions":{"x-order":"2147483647"}}},"/blog/of/user":{"get":{"tags":["blog-controller"],"summary":"queryBlogByUserId","operationId":"queryBlogByUserIdUsingGET","parameters":[{"name":"current","in":"query","description":"current","required":false,"style":"form","schema":{"type":"integer","format":"int32"}},{"name":"id","in":"query","description":"id","required":true,"style":"form","schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Result"}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"extensions":{"x-order":"2147483647"}}},"/blog/{id}":{"get":{"tags":["blog-controller"],"summary":"queryBlogById","operationId":"queryBlogByIdUsingGET","parameters":[{"name":"id","in":"path","description":"id","required":true,"style":"simple","schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Result"}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"extensions":{"x-order":"2147483647"}}},"/follow/common/{id}":{"get":{"tags":["follow-controller"],"summary":"common","operationId":"commonUsingGET","parameters":[{"name":"id","in":"path","description":"id","required":true,"style":"simple","schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Result"}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"extensions":{"x-order":"2147483647"}}},"/follow/or/not/{id}":{"get":{"tags":["follow-controller"],"summary":"isFollow","operationId":"isFollowUsingGET","parameters":[{"name":"id","in":"path","description":"id","required":true,"style":"simple","schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Result"}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"extensions":{"x-order":"2147483647"}}},"/follow/{id}/{isfollow}":{"put":{"tags":["follow-controller"],"summary":"follow","operationId":"followUsingPUT","parameters":[{"name":"id","in":"path","description":"id","required":true,"style":"simple","schema":{"type":"integer","format":"int64"}},{"name":"isfollow","in":"path","description":"isfollow","required":true,"style":"simple","schema":{"type":"boolean"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Result"}}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"extensions":{"x-order":"2147483647"}}},"/shop":{"put":{"tags":["shop-controller"],"summary":"updateShop","operationId":"updateShopUsingPUT","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Shop"}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Result"}}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"extensions":{"x-order":"2147483647"}},"post":{"tags":["shop-controller"],"summary":"saveShop","operationId":"saveShopUsingPOST","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Shop"}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Result"}}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"extensions":{"x-order":"2147483647"}}},"/shop/of/name":{"get":{"tags":["shop-controller"],"summary":"queryShopByName","operationId":"queryShopByNameUsingGET","parameters":[{"name":"name","in":"query","description":"name","required":false,"style":"form","schema":{"type":"string"}},{"name":"current","in":"query","description":"current","required":false,"style":"form","schema":{"type":"integer","format":"int32"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Result"}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"extensions":{"x-order":"2147483647"}}},"/shop/of/type":{"get":{"tags":["shop-controller"],"summary":"queryShopByType","operationId":"queryShopByTypeUsingGET","parameters":[{"name":"typeId","in":"query","description":"typeId","required":true,"style":"form","schema":{"type":"integer","format":"int32"}},{"name":"current","in":"query","description":"current","required":false,"style":"form","schema":{"type":"integer","format":"int32"}},{"name":"x","in":"query","description":"x","required":false,"style":"form","schema":{"type":"number","format":"double"}},{"name":"y","in":"query","description":"y","required":false,"style":"form","schema":{"type":"number","format":"double"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Result"}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"extensions":{"x-order":"2147483647"}}},"/shop/{id}":{"get":{"tags":["shop-controller"],"summary":"queryShopById","operationId":"queryShopByIdUsingGET","parameters":[{"name":"id","in":"path","description":"id","required":true,"style":"simple","schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Result"}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"extensions":{"x-order":"2147483647"}}},"/shop-type/list":{"get":{"tags":["shop-type-controller"],"summary":"queryTypeList","operationId":"queryTypeListUsingGET","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Result"}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"extensions":{"x-order":"2147483647"}}},"/upload/blog":{"post":{"tags":["upload-controller"],"summary":"uploadImage","operationId":"uploadImageUsingPOST","requestBody":{"content":{"application/json":{"schema":{"type":"string","format":"binary"}},"application/octet-stream":{"schema":{"type":"string","format":"binary"}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Result"}}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"extensions":{"x-order":"2147483647"}}},"/upload/blog/delete":{"get":{"tags":["upload-controller"],"summary":"deleteBlogImg","operationId":"deleteBlogImgUsingGET","parameters":[{"name":"name","in":"query","description":"name","required":true,"style":"form","schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Result"}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"extensions":{"x-order":"2147483647"}}},"/user/code":{"post":{"tags":["user-controller"],"summary":"sendCode","operationId":"sendCodeUsingPOST","parameters":[{"name":"creationTime","in":"query","required":false,"style":"form","schema":{"type":"integer","format":"int64"}},{"name":"id","in":"query","required":false,"style":"form","schema":{"type":"string"}},{"name":"lastAccessedTime","in":"query","required":false,"style":"form","schema":{"type":"integer","format":"int64"}},{"name":"maxInactiveInterval","in":"query","required":false,"style":"form","schema":{"type":"integer","format":"int32"}},{"name":"new","in":"query","required":false,"style":"form","schema":{"type":"boolean"}},{"name":"phone","in":"query","description":"phone","required":true,"style":"form","schema":{"type":"string"}},{"name":"valueNames","in":"query","required":false,"style":"pipeDelimited","schema":{"type":"array","items":{"type":"string"}}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Result"}}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"extensions":{"x-order":"2147483647"}}},"/user/info/{id}":{"get":{"tags":["user-controller"],"summary":"info","operationId":"infoUsingGET","parameters":[{"name":"id","in":"path","description":"id","required":true,"style":"simple","schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Result"}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"extensions":{"x-order":"2147483647"}}},"/user/login":{"post":{"tags":["user-controller"],"summary":"login","operationId":"loginUsingPOST","parameters":[{"name":"creationTime","in":"query","required":false,"style":"form","schema":{"type":"integer","format":"int64"}},{"name":"id","in":"query","required":false,"style":"form","schema":{"type":"string"}},{"name":"lastAccessedTime","in":"query","required":false,"style":"form","schema":{"type":"integer","format":"int64"}},{"name":"maxInactiveInterval","in":"query","required":false,"style":"form","schema":{"type":"integer","format":"int32"}},{"name":"new","in":"query","required":false,"style":"form","schema":{"type":"boolean"}},{"name":"valueNames","in":"query","required":false,"style":"pipeDelimited","schema":{"type":"array","items":{"type":"string"}}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginFormDTO"}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Result"}}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"extensions":{"x-order":"2147483647"}}},"/user/logout":{"post":{"tags":["user-controller"],"summary":"logout","operationId":"logoutUsingPOST","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Result"}}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"extensions":{"x-order":"2147483647"}}},"/user/me":{"get":{"tags":["user-controller"],"summary":"me","operationId":"meUsingGET","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Result"}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"extensions":{"x-order":"2147483647"}}},"/user/sign":{"post":{"tags":["user-controller"],"summary":"sign","operationId":"signUsingPOST","responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Result"}}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"extensions":{"x-order":"2147483647"}}},"/user/{id}":{"get":{"tags":["user-controller"],"summary":"queryUserById","operationId":"queryUserByIdUsingGET","parameters":[{"name":"id","in":"path","description":"id","required":true,"style":"simple","schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Result"}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"extensions":{"x-order":"2147483647"}}},"/voucher":{"post":{"tags":["voucher-controller"],"summary":"addVoucher","operationId":"addVoucherUsingPOST","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Voucher"}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Result"}}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"extensions":{"x-order":"2147483647"}}},"/voucher/list/{shopId}":{"get":{"tags":["voucher-controller"],"summary":"queryVoucherOfShop","operationId":"queryVoucherOfShopUsingGET","parameters":[{"name":"shopId","in":"path","description":"shopId","required":true,"style":"simple","schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Result"}}}},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"extensions":{"x-order":"2147483647"}}},"/voucher/seckill":{"post":{"tags":["voucher-controller"],"summary":"addSeckillVoucher","operationId":"addSeckillVoucherUsingPOST","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/Voucher"}}}},"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Result"}}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"extensions":{"x-order":"2147483647"}}},"/voucher-order/seckill/{id}":{"post":{"tags":["voucher-order-controller"],"summary":"seckillVoucher","operationId":"seckillVoucherUsingPOST","parameters":[{"name":"id","in":"path","description":"id","required":true,"style":"simple","schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"*/*":{"schema":{"$ref":"#/components/schemas/Result"}}}},"201":{"description":"Created"},"401":{"description":"Unauthorized"},"403":{"description":"Forbidden"},"404":{"description":"Not Found"}},"extensions":{"x-order":"2147483647"}}}},"components":{"schemas":{"Blog":{"title":"Blog","type":"object","properties":{"comments":{"type":"integer","format":"int32"},"content":{"type":"string"},"createTime":{"type":"string","format":"date-time"},"icon":{"type":"string"},"id":{"type":"integer","format":"int64"},"images":{"type":"string"},"isLike":{"type":"boolean"},"liked":{"type":"integer","format":"int32"},"name":{"type":"string"},"shopId":{"type":"integer","format":"int64"},"title":{"type":"string"},"updateTime":{"type":"string","format":"date-time"},"userId":{"type":"integer","format":"int64"}}},"LoginFormDTO":{"title":"LoginFormDTO","type":"object","properties":{"code":{"type":"string"},"password":{"type":"string"},"phone":{"type":"string"}}},"Result":{"title":"Result","type":"object","properties":{"data":{"type":"object"},"errorMsg":{"type":"string"},"success":{"type":"boolean"},"total":{"type":"integer","format":"int64"}}},"Shop":{"title":"Shop","type":"object","properties":{"address":{"type":"string"},"area":{"type":"string"},"avgPrice":{"type":"integer","format":"int64"},"comments":{"type":"integer","format":"int32"},"createTime":{"type":"string","format":"date-time"},"distance":{"type":"number","format":"double"},"id":{"type":"integer","format":"int64"},"images":{"type":"string"},"name":{"type":"string"},"openHours":{"type":"string"},"score":{"type":"integer","format":"int32"},"sold":{"type":"integer","format":"int32"},"typeId":{"type":"integer","format":"int64"},"updateTime":{"type":"string","format":"date-time"},"x":{"type":"number","format":"double"},"y":{"type":"number","format":"double"}}},"Voucher":{"title":"Voucher","type":"object","properties":{"actualValue":{"type":"integer","format":"int64"},"beginTime":{"type":"string","format":"date-time"},"createTime":{"type":"string","format":"date-time"},"endTime":{"type":"string","format":"date-time"},"id":{"type":"integer","format":"int64"},"payValue":{"type":"integer","format":"int64"},"rules":{"type":"string"},"shopId":{"type":"integer","format":"int64"},"status":{"type":"integer","format":"int32"},"stock":{"type":"integer","format":"int32"},"subTitle":{"type":"string"},"title":{"type":"string"},"type":{"type":"integer","format":"int32"},"updateTime":{"type":"string","format":"date-time"}}}}}}